Revive node:http tests
